v0.2.0

+++ src/app/features/board/components/navbar/navbar.component.ts
@@ -8,11 +8,17 @@ import { BrowserModule } from '@angular/platform-browser';
 import {MatMenuModule} from '@angular/material/menu';
 import {MatButtonModule} from '@angular/material/button';
 import { IconService } from '@shared-services/icon/icon.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteConfirmationComponent } from '../../../account/components/edit-board-modal/components/delete-confirmation/delete-confirmation.component';
+import { MatTooltip } from '@angular/material/tooltip';
+import { SettingsModalComponent } from '../../../account/components/sidebar/components/modals/settings-modal/settings-modal.component';
+import download from '@util-functions/download';
+import {MatSnackBar} from '@angular/material/snack-bar';
 
 @Component({
   selector: 'navbar-component',
   standalone: true,
-  imports: [RouterModule, MatIconModule, MatButtonModule, MatMenuModule],
+  imports: [RouterModule, MatIconModule, MatButtonModule, MatMenuModule, MatTooltip],
   templateUrl: './navbar.component.html',
   styleUrl: './navbar.component.scss'
 })
@@ -27,7 +33,9 @@ export class NavbarComponent {
     private renderer: Renderer2,
     private activatedRoute: ActivatedRoute,
     private router: Router,
-    private iconService: IconService
+    private iconService: IconService,
+    protected dialog: MatDialog,
+    protected snackBar: MatSnackBar
   ) {
 
   }
@@ -40,16 +48,53 @@ export class NavbarComponent {
 
   }
 
-  deleteBoard() {
-    //!Confirmation
-    this.boardData.deleteBoard(this.boardData.activeId)
-    this.router.navigate(['/account'])
-  }
-
   saveData() {
     this.boardData.saveData();
     this.router.navigate(['/account'])
   }
 
+  confirmDelete() {
+    const dialog = this.dialog.open(DeleteConfirmationComponent)
+
+    dialog.afterClosed().subscribe((result)=>{
+      if(result==="DELETE") {
+        this.boardData.deleteBoard(this.boardData.activeId);
+        this.router.navigate(['/account']);
+      }
+    })
+  }
+
+  settingsModal() {
+    const modalCookies = this.dialog.open(SettingsModalComponent);
+
+    modalCookies.afterClosed().subscribe((result)=>{
+    })
+  }
+
+  downloadBoard() {
+    const board = this.boardData.getActiveBoard()
+    if(board)
+      download(JSON.stringify(board),board.name,'application/json','json')
+  }
+
+  saveAndDuplicate() {
+    let board = this.boardData.getActiveBoard()
+    if(!board) return;
+    board.name = `${board.name} duplicate`
+    this.boardData.saveData();
+    this.boardData.createBoard(board,false);
+    this.openSnackBar('Board created','Ok');
+  }
+
+  saveAndCreate() {
+    this.boardData.saveData();
+    this.boardData.createBoard(undefined,true)
+  }
+
+  openSnackBar(message: string, action: string) {
+    this.snackBar.open(message, action, {
+      duration: 5000
+    });
+  }
 
 }
